pyproject.toml
[tool.poetry]
name = "srp-file-processor"
version = "0.1.0"
description = "Processes CSV files exported from the Statistical Report Program (SRP)"
authors = ["Stephen Fuqua <stephen@safnet.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.3"
SQLAlchemy = "^1.3.20"
python-dotenv = "^0.14.0"
xlrd3 = "^1.0.0"
dash = "^1.16.3"
ipython = "^7.19.0"
ipykernel = "^5.4.2"
jupyter = "^1.0.0"
pandoc = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.790"
pytest-describe = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
